Permettre nombre à virgule pour le rpix

+++ database/migrations/1638413722616_billet_changes.ts
@@ -18,7 +18,7 @@ export default class BilletChanges extends BaseSchema {
     this.schema.createTable(this.tableName, (table) => {
       table.increments('id')
       table.string('nom').defaultTo('Unnamed')
-      table.integer('prix').defaultTo(10)
+      table.float('prix').defaultTo(10)
       table.boolean('extra').defaultTo(false)
       table.timestamp('created_at', { useTz: true }).defaultTo(this.now())
       table.timestamp('updated_at', { useTz: true }).defaultTo(this.now())
